requirements.txt
numpy==1.21.5
Pillow==9.0.1
Pillow==10.3.0
